requirements.txt
dj-database-url==0.4.2
Django==1.11.7
django-filter==1.1.0
django-pagedown==0.1.3
djangorestframework==3.7.3
Markdown==2.6.9
olefile==0.44
Pillow==4.3.0
python-decouple==3.1
pytz==2017.3
